+++ src/modules/ai/config/openai.config.ts
@@ -1,14 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
 import OpenAI from 'openai';
 
 @Injectable()
 export class OpenAiConfig {
   private readonly openai: OpenAI;
 
-  constructor(private readonly configService: ConfigService) {
-    const apiKey = this.configService.get<string>('OPEN_AI');
-    const projectId = this.configService.get<string>('PROJECT_ID'); // opcional
+  constructor() {
+    const apiKey = process.env.OPEN_AI;
+    const projectId = process.env.PROJECT_ID; // opcional
 
     this.openai = new OpenAI({
       apiKey,