+++ src/modules/ai/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './active-intent.utils';
 export * from './serialize-context.utils';
 export * from './formated-date.utils';
+export * from './parse-response';